moved AverageRatings interface to Rating.ts
