--- frontend/src/interfaces/AverageRatings.ts
@@ -1,9 +1,0 @@
-export interface AverageRatings {
-    amount: number;
-    totalRating: number;
-    punctuality: number;
-    reliability: number;
-    comfortDuringTrip: number;
-    cargoArrivedUndamaged: number;
-    passengerPleasantness: number;
-}
